+++ src/lib.rs
@@ -15,6 +15,7 @@
 //!     IntId,
 //!     gicv3::{
 //!         GicV3, SgiTarget, SgiTargetGroup,
+//!         cpu_interface::GicCpuInterface,
 //!         registers::{Gicd, GicrSgi},
 //!     },
 //!     irq_enable,
@@ -30,11 +31,11 @@
 //!
 //! // Configure an SGI and then send it to ourself.
 //! let sgi_intid = IntId::sgi(3);
-//! GicV3::set_priority_mask(0xff);
+//! GicCpuInterface::set_priority_mask(0xff);
 //! gic.set_interrupt_priority(sgi_intid, Some(0), 0x80);
 //! gic.enable_interrupt(sgi_intid, Some(0), true);
 //! irq_enable();
-//! GicV3::send_sgi(
+//! GicCpuInterface::send_sgi(
 //!     sgi_intid,
 //!     SgiTarget::List {
 //!         affinity3: 0,
@@ -183,9 +184,36 @@ impl IntId {
         Self::PPI_START <= self.0 && self.0 < Self::SPI_START
     }
 
-    /// Returns whether this interrupt ID is private to a core, i.e. it is an SGI or PPI.
+    /// Returns whether this interrupt ID is for an Extended Private Peripheral Interrupt.
+    pub const fn is_eppi(self) -> bool {
+        Self::EPPI_START <= self.0 && self.0 < Self::EPPI_END
+    }
+
+    /// Returns whether this interrupt ID is private to a core, i.e. it is an SGI, PPI or EPPI.
     pub const fn is_private(self) -> bool {
-        self.is_sgi() || self.is_ppi()
+        self.is_sgi() || self.is_ppi() || self.is_eppi()
+    }
+
+    /// Returns SGI index or `None` if it is not an SGI interrupt ID.
+    pub const fn sgi_index(self) -> Option<u32> {
+        if self.is_sgi() {
+            Some(self.0 - Self::SGI_START)
+        } else {
+            None
+        }
+    }
+
+    /// Maps SGI, PPI and EPPI interrupt IDs into a continuous index range starting from 0,
+    /// making it ideal for indexing redistributor registers.
+    pub const fn private_index(self) -> Option<usize> {
+        if self.is_sgi() || self.is_ppi() {
+            Some(self.0 as usize)
+        } else if self.is_eppi() {
+            // EPPI
+            Some((self.0 - IntId::EPPI_START + IntId::SGI_COUNT + IntId::PPI_COUNT) as usize)
+        } else {
+            None
+        }
     }
 
     /// Returns whether this interrupt ID is for a Shared Peripheral Interrupt.